pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "My second education project in online school HEXLET.IO (level 2)"
repository = "https://github.com/alexartoff/python-project-lvl2"
authors = ["alexandertarakanov <alex.artoff@gmail.com>"]
packages = [{include = "gendiff" },]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
